Fixed Non-Slow Snake Block Crash (read desc)
When the speed of the snake block wasn't set to "very slow", the game would crash after the snake block moved a few blocks away.

That seemed to be caused by the speed value interfering with the daEnSnakeBlock_c.someValueCalculatedFromAnmClrAndSpeed value and caused the game try to apply the "create" CLR0 animation more than it should have.

It was fixed by capping the amount of time of this animation to be applied to 2 like it does with the very slow speed.

This is pretty hacky so consider this fix as temporary