Fixed Tile God Bug
This bug made it so disabling SFXs and GFXs from a tilegod sprite would also made it place its tiles in a checkers pattern.

This is due to the pattern value also containing the SFX/GFX flags and the onCreate function from the top of the file not actually being hooked into the game.

Fixed it by making a new variable named actualPattern that actually contains the pattern that is being used.